0.2.1

+++ src/timeline_list.vala
@@ -107,7 +107,7 @@ public class TimelineList : TimelineListAbstract {
 			return result;
 		}
 		
-		if(result.size > 0) { //if we get some statuses
+		if(result.size > 0) { //if we got some statuses
 			if((!_parent_focus || !act.active) && lst.size > 0) { //if this list is not visible and we have updates
 				
 				have_fresh = true;
@@ -124,7 +124,9 @@ public class TimelineList : TimelineListAbstract {
 			}
 		}
 		
-		delete_extra();
+		if(!parent_focus) //delete only if main window out of focus
+			delete_extra();
+		
 		warning("SIZE: %d", lst.size);
 		
 		refresh();
